requirements.txt
pydantic
httpx
websockets
python-dotenv